Add time since last physical interaction and idle callbacks
This adds some tooling which keeps track of the time elapsed since
the last time the user interacted with the keyboard in any way.
The primary goal of this is to make it easier to disable hardware
features when the user is not using the keyboard, lowering power
consumption and preventing burn-in.